Solve softmax-function

+++ softmax-function/softmax-function.py
@@ -4,7 +4,7 @@ def softmax(x: list) -> np.ndarray:
     """
     Returns stable softmax probabilities as a NumPy array matching the shape of x.
     """
-    x = np.asarray(x)
+    x = np.asarray(x, dtype = float)
     m = np.max(x, axis = -1, keepdims=True)
 
     x = x-m
